refactor: modify file structor
