+++ packages/vtable-plugins/src/index.ts
@@ -1,3 +1,4 @@
 export * from './carousel-animation';
 export * from './invert-highlight';
 export * from './header-highlight';
+export * from './plus-row-column';
