+++ packages/vtable-plugins/src/plus-row-column.ts
@@ -1,0 +1,10 @@
+import type { BaseTableAPI } from '@visactor/vtable/src/ts-types/base-table';
+
+export class PlusRowColumnPlugin {
+  constructor(private table: BaseTableAPI) {
+    this.table = table;
+  }
+  init() {
+    //
+  }
+}
