+++ packages/vtable/src/chartModule.ts
@@ -1,7 +1,7 @@
 /* eslint-disable sort-imports */
 import { extend } from './tools/helper';
-import { chartTypes as plugins } from './plugins/chartModules';
+export const chartTypes: { [key: string]: any } = {};
 const builtin = {};
 export function get(): { [key: string]: any } {
-  return extend(builtin, plugins);
+  return extend(builtin, chartTypes);
 }
