--- packages/vtable/src/plugins/chartModules.ts
@@ -1,1 +1,0 @@
-export const chartTypes: { [key: string]: any } = {};
