+++ packages/vtable/src/themes.ts
@@ -5,8 +5,8 @@ import brightTheme from './themes/BRIGHT';
 import arcoTheme from './themes/ARCO';
 import defaultTheme from './themes/DEFAULT';
 import materialDesignTheme from './themes/SIMPLIFY';
-import { themes as plugins } from './plugins/themes';
-import { TableTheme } from './themes/theme';
+import { themes as plugins } from './themes/themes';
+import { TableTheme } from './themes/theme-define';
 import type { ITableThemeDefine } from './ts-types';
 export const DARK = new TableTheme(darkTheme, darkTheme);
 export const BRIGHT = new TableTheme(brightTheme, brightTheme);
