+++ packages/vtable/src/themes/themes.ts
@@ -1,3 +1,3 @@
-import type { TableTheme } from '../themes/theme';
+import type { TableTheme } from '../themes/theme-define';
 
 export const themes: { [key: string]: TableTheme } = {};
