+++ packages/vtable/src/ts-types/table-engine.ts
@@ -33,7 +33,7 @@ import type { EditManager } from '../edit/edit-manager';
 import type { ICustomRender } from './customElement';
 import type { ICustomLayout } from './customLayout';
 import type { ColorPropertyDefine, StylePropertyFunctionArg } from './style-define';
-import type { TableTheme } from '../themes/theme';
+import type { TableTheme } from '../themes/theme-define';
 
 export interface CellAddress {
   col: number;
